Initial commit, bare bones structure of server with some logic.
This javascript code uses an  arithmatic formula to return answer to user, based on the variable that user provides.
only admins can modify the constants used in the formula and each modification log is stored in database.
User can only send the variable and ask for an answer in return.